requirements.txt
requests
python-dotenv
pandas
sqlalchemy
psycopg2-binary 